🧐 Forenkling av useActivePeriod hook
