+++ src/app/person/[aktorId]/error.module.scss
@@ -1,0 +1,6 @@
+.error {
+    border-radius: 0;
+    height: max-content;
+    grid-column-start: venstremeny;
+    grid-column-end: høyremeny;
+}
