+++ src/app/person/[aktorId]/inngangsvilk%C3%A5r/page.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react';
 
-import PeriodeView from '@/app/person/[aktorId]/periodeview';
+import PeriodeView from '@/routes/saksbilde/periodeview/PeriodeView';
 
 function Page(): ReactElement | null {
     return <PeriodeView />;
