--- src/app/person/[aktorId]/periodeview.module.scss
@@ -1,15 +1,0 @@
-.content {
-  grid-area: content;
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  height: 100%;
-  min-width: 0;
-}
-
-.error {
-  border-radius: 0;
-  height: max-content;
-  grid-column-start: venstremeny;
-  grid-column-end: høyremeny;
-}
